Store Promotions: Move fixture
The fixtures.js file caused an error with Jest, so this moves it to a
fixtures/ folder to resolve it.
