+++ client/extensions/woocommerce/state/sites/promotions/test/fixtures/promotions.js
